Disable debug mode by default
