+++ debug.go
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2014 Christian Muehlhaeuser
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ *	Authors:
+ *		Christian Muehlhaeuser <muesli@gmail.com>
+ *		Michael Wendland <michael@michiwend.com>
+ */
+
+/*
+Package smartcrop implements a content aware image cropping library based on
+Jonas Wagner's smartcrop.js https://github.com/jwagner/smartcrop.js
+*/
+package smartcrop
+
+import (
+	"image"
+	"image/jpeg"
+	"image/png"
+	"os"
+)
+
+func debugOutput(debug bool, img *image.Image, debugType string) {
+	if debug {
+		writeImageToPng(img, "./smartcrop_"+debugType+".png")
+	}
+}
+
+func writeImageToJpeg(img *image.Image, name string) {
+	fso, err := os.Create(name)
+	if err != nil {
+		panic(err)
+	}
+	defer fso.Close()
+
+	jpeg.Encode(fso, (*img), &jpeg.Options{Quality: 100})
+}
+
+func writeImageToPng(img *image.Image, name string) {
+	fso, err := os.Create(name)
+	if err != nil {
+		panic(err)
+	}
+	defer fso.Close()
+
+	png.Encode(fso, (*img))
+}
